+++ src/main/java/com/sudoplay/joise/noise/worker/WorkerNoise2Value.java
@@ -50,7 +50,7 @@ package com.sudoplay.joise.noise.worker;
 
 import com.sudoplay.joise.noise.Noise;
 import com.sudoplay.joise.noise.worker.spi.WorkerNoise2;
-import com.sudoplay.util.Bits;
+import com.sudoplay.joise.util.Bits;
 
 public class WorkerNoise2Value implements
     WorkerNoise2 {